wip: refactoring input into drag
