+++ src/input/hover.rs
@@ -1,0 +1,14 @@
+use crate::prelude::*;
+
+
+pub(crate) struct HoverPlugin;
+
+impl Plugin for HoverPlugin {
+    fn build(&self, app: &mut App) {
+        
+    }
+}
+
+fn hover() {
+    
+}
